+++ extensions/cronet/src/main/java/com/google/android/exoplayer2/ext/cronet/CronetDataSource.java
@@ -140,13 +140,6 @@ public class CronetDataSource extends BaseDataSource implements HttpDataSource {
       readTimeoutMs = DEFAULT_READ_TIMEOUT_MILLIS;
     }
 
-    /** @deprecated Use {@link #setDefaultRequestProperties(Map)} instead. */
-    @Deprecated
-    @Override
-    public final RequestProperties getDefaultRequestProperties() {
-      return defaultRequestProperties;
-    }
-
     @Override
     public final Factory setDefaultRequestProperties(Map<String, String> defaultRequestProperties) {
       this.defaultRequestProperties.clearAndSet(defaultRequestProperties);